CHAD-17040: zwave-thermostat: Subdriver modifications for lazy loading
