+++ drivers/SmartThings/zwave-thermostat/src/aeotec-radiator-thermostat/init.lua
@@ -1,16 +1,6 @@
--- Copyright 2022 SmartThings
---
--- Licensed under the Apache License, Version 2.0 (the "License");
--- you may not use this file except in compliance with the License.
--- You may obtain a copy of the License at
---
---     http://www.apache.org/licenses/LICENSE-2.0
---
--- Unless required by applicable law or agreed to in writing, software
--- distributed under the License is distributed on an "AS IS" BASIS,
--- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
--- See the License for the specific language governing permissions and
--- limitations under the License.
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
 
 local capabilities = require "st.capabilities"
 --- @type st.zwave.CommandClass
@@ -23,12 +13,6 @@ local Battery = (require "st.zwave.CommandClass.Battery")({version=1})
 local SensorMultilevel = (require "st.zwave.CommandClass.SensorMultilevel")({version=2})
 --- @type st.zwave.CommandClass.ThermostatSetpoint
 local ThermostatSetpoint = (require "st.zwave.CommandClass.ThermostatSetpoint")({version=1})
-
-local AEOTEC_THERMOSTAT_FINGERPRINT = {mfr = 0x0371, prod = 0x0002, model = 0x0015}
-
-local function can_handle_aeotec_radiator_thermostat(opts, driver, device, ...)
-  return device:id_match(AEOTEC_THERMOSTAT_FINGERPRINT.mfr, AEOTEC_THERMOSTAT_FINGERPRINT.prod, AEOTEC_THERMOSTAT_FINGERPRINT.model)
-end
 
 local function thermostat_mode_report_handler(self, device, cmd)
   local event = nil
@@ -114,7 +98,7 @@ local aeotec_radiator_thermostat = {
     }
   },
   lifecycle_handlers = {added = device_added},
-  can_handle = can_handle_aeotec_radiator_thermostat
+  can_handle = require("aeotec-radiator-thermostat.can_handle"),
 }
 
 return aeotec_radiator_thermostat
