+++ drivers/SmartThings/zwave-thermostat/src/ct100-thermostat/fingerprints.lua
@@ -1,0 +1,9 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local CT100_THERMOSTAT_FINGERPRINTS = {
+  { manufacturerId = 0x0098, productType = 0x6401, productId = 0x0107 }, -- 2Gig CT100 Programmable Thermostat
+  { manufacturerId = 0x0098, productType = 0x6501, productId = 0x000C }, -- Iris Thermostat
+}
+
+return CT100_THERMOSTAT_FINGERPRINTS
