+++ drivers/SmartThings/zwave-thermostat/src/fibaro-heat-controller/fingerprints.lua
@@ -1,0 +1,9 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local FIBARO_HEAT_FINGERPRINTS = {
+    {mfr = 0x010F, prod = 0x1301, model = 0x1000}, -- Fibaro Heat Controller
+    {mfr = 0x010F, prod = 0x1301, model = 0x1001} -- Fibaro Heat Controller
+}
+
+return FIBARO_HEAT_FINGERPRINTS
