+++ drivers/SmartThings/zwave-thermostat/src/qubino-flush-thermostat/can_handle.lua
@@ -1,0 +1,14 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local function can_handle_qubino_thermostat(opts, driver, device, ...)
+  local FINGERPRINTS = require("qubino-flush-thermostat.fingerprints")
+  for _, fingerprint in ipairs(FINGERPRINTS) do
+    if device:id_match(fingerprint.mfr, fingerprint.prod, fingerprint.model) then
+      return true, require("qubino-flush-thermostat")
+    end
+  end
+  return false
+end
+
+return can_handle_qubino_thermostat
