+++ drivers/SmartThings/zwave-thermostat/src/qubino-flush-thermostat/fingerprints.lua
@@ -1,0 +1,8 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local QUBINO_FINGERPRINTS = {
+  {mfr = 0x0159, prod = 0x0005, model = 0x0054},  -- Qubino Flush On/Off Thermostat 2
+}
+
+return QUBINO_FINGERPRINTS
