+++ drivers/SmartThings/zwave-thermostat/src/stelpro-ki-thermostat/fingerprints.lua
@@ -1,0 +1,8 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local STELPRO_KI_THERMOSTAT_FINGERPRINTS = {
+  { manufacturerId = 0x0239, productType = 0x0001, productId = 0x0001 } -- Stelpro Ki Thermostat
+}
+
+return STELPRO_KI_THERMOSTAT_FINGERPRINTS
