+++ drivers/SmartThings/zwave-thermostat/src/thermostat-heating-battery/fingerprints.lua
@@ -1,0 +1,9 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local DANFOSS_LC13_THERMOSTAT_FINGERPRINTS = {
+    { manufacturerId = 0x0002, productType = 0x0005, productId = 0x0003 }, -- Danfoss LC13 Thermostat
+    { manufacturerId = 0x0002, productType = 0x0005, productId = 0x0004 } -- Danfoss LC13 Thermostat
+}
+
+return DANFOSS_LC13_THERMOSTAT_FINGERPRINTS
